refactor(shared): unify model param-count parsing across stack
The backend and frontend each carried their own parameter-count parser that had
drifted (whitespace boundaries, the `illion` suffix, and the backend-only `<
10000` sanity guard), so a fix to one never reached the other. Consolidate on a
single source of truth.

- add canonical `parseParameterCountFromName` and `resolveModelParamCount` to
  `@airunway/shared` (`shared/types/modelParams.ts`), keeping the stricter
  backend parsing behaviour
- rewire `modelCompatibility.ts` to import `parseParameterCountFromName` from
  shared, and drop the now-dead `resolveParamCount` from `gpuPerformance.ts` (no
  production callers remained)
- replace the divergent inline regex in `gpu-throughput-params.ts` with the
  shared `resolveModelParamCount`
- repoint the backend tests to the shared implementation
