+++ shared/types/index.ts
@@ -1,4 +1,5 @@
 export * from './model';
+export * from './modelParams';
 export * from './deployment';
 export * from './settings';
 export * from './installation';
